LayoutTests/css3/filters/custom-with-at-rule-syntax/parsing-at-rule-invalid-expected.txt

Test at-rule parsing for @-webkit-filter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



========================================
Filter at-rule symbol tests.
========================================

Unprefixed rule.
@filter my-filter { }
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

========================================
Filter name tests.
========================================

Missing filter name identifier.
@-webkit-filter { }
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

Filter name as string.
@-webkit-filter 'my-filter' { }
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

Filter name as number.
@-webkit-filter 123 { }
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

========================================
Filter at-rule body tests.
========================================

Missing rule body.
@-webkit-filter my-filter
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

Missing opening brace.
@-webkit-filter my-filter }
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR

Missing closing brace.
@-webkit-filter my-filter {
PASS insertRuleException instanceof DOMException is true
PASS insertRuleException.code is DOMException.SYNTAX_ERR
PASS successfullyParsed is true

TEST COMPLETE
